refactor: moved global folders to shared for better clarity
