+++ src/features/document-listing/components/DocumentCard.tsx
@@ -1,7 +1,7 @@
-import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import { timeFromNow } from "@/lib/datetime";
-import { cn } from "@/lib/utils";
-import type { Document } from "@/types/document";
+import { Card, CardContent, CardHeader } from "@/shared/components/ui/card";
+import { timeFromNow } from "@/shared/lib/datetime";
+import { cn } from "@/shared/lib/utils";
+import type { Document } from "@/shared/types/document";
 import { Clock } from "lucide-react";
 import Image from "next/image";
 import React from "react";
