+++ src/shared/components/common/ResumeCraftBrand.tsx
