+++ src/shared/components/common/ResumeCraftIcon.tsx
