+++ src/shared/types/document.ts
